Add dependencies to .mod file
Tentatively we could do away with the whole load order thing. But I
don't know how many HFM off-shoots there are out there, and I can't
cover them all in the limited time I am willing to spend on this last
release.

As a consequence the dependency list is more or less a last-ditch
measure to ensure good load order.